Make category codes selection a dropdown
