+++ components/aggregation/mat/category_codes.json
@@ -1,0 +1,32 @@
+[
+  "ALDR",
+  "ANON",
+  "COMM",
+  "COMT",
+  "CTRL",
+  "CULTR",
+  "DATE",
+  "ECON",
+  "ENV",
+  "FILE",
+  "GAME",
+  "GMB",
+  "GOVT",
+  "GRP",
+  "HACK",
+  "HATE",
+  "HOST",
+  "HUMR",
+  "IGO",
+  "LGBT",
+  "MILX",
+  "MMED",
+  "NEWS",
+  "POLR",
+  "PORN",
+  "PROV",
+  "PUBH",
+  "REL",
+  "SRCH",
+  "XED"
+]
